[FEATURE] Initial commit of nabi.m.ImageFileUploader
Attention - this is Work-In-Progress
Currently, this control uses private APIs of the unified FileUploader.

There is much room for other improvements, i.e. in the grunt file. See also all the comments in the code. But the most important thing is to make the unified FileUploader offer some better APIs, i.e. @interface APIs. I will open a pull request for this soon.